requirements.txt
 numpy
 pandas
 matplotlib
 andas_datareader
 keras.models
load_model
streamlit
seaborn